requirements.txt
fastapi
uvicorn
replicate
python-multipart
python-dotenv
requests
asgiref